refactor(l5): move operator-doctor from src/ into server/application behind an L5 adapter (issue 864)
src/operator-doctor.ts was 770 lines and read twelve env keys inline, so it
could neither meet the server/ file rule nor the server/ no-process-env rule as
a single file. It moves under server/application/ split into three siblings by
responsibility, and the old src/ path becomes an M9 adapter so every legacy
caller keeps its call form unchanged.

Moved, old -> new:

- src/operator-doctor.ts -> server/application/operator-doctor.ts
  (assembly of the frozen payload + the single-flight short-TTL cache)
- new sibling server/application/operator-doctor-types.ts
  (the frozen OperatorDoctorStatus shape, DOCTOR_CONTRACT_VERSION, the
  distillation-lag ratio constants, canReadDoctor, classifyDistillationLag,
  and the widened OperatorDoctorOptions)
- new sibling server/application/operator-doctor-probes.ts
  (the individual diagnostic questions: service version, migrations, audit
  storage, distillation lag, embedding availability, qmd binary and index
  probe worker, withTimeout and probeUrl)

Every sibling satisfies the 500-line file rule and the 100-line function rule,
and none of the three reads process.env.

The twelve env reads become fields of ONE options object. The existing 4th
parameter type OperatorDoctorCacheOptions widened into OperatorDoctorOptions
with the env-derived fields required there, so the parameter count stays 4 and
getOperatorDoctorStatus(pool, natsRuntimeBoundary, natsBridgeHealth, options)
keeps its signature. Where server/main.ts:403 gets each field:

- serviceVersionFallback  <- config.doctor.serviceVersionFallback
- nodeEnvironment         <- config.doctor.nodeEnvironment
- fileLogConfigured       <- config.logging.file (already parsed)
- rotationConfigured      <- config.doctor.rotationConfigured
- rawTurnTtlSeconds       <- config.doctor.rawTurnTtlSeconds
- qmdIndexPath            <- config.doctor.qmdIndexPath (spread when present)

Two of those had no reader yet, so server/config/src-readers.ts gains two
one-line raw-optional fields and the doctorGroup lines that shape them:
OPENBRAIN_RAW_TURN_TTL_SECONDS (the distillation-lag denominator, reproducing
readDistillationLagTtlSeconds exactly -- Number(), not parseInt, so a trailing
suffix falls back) and npm_package_version (the version reported when
package.json itself cannot be read).

The adapter at src/operator-doctor.ts sits under the M9 header, imports only
from server/ paths, and re-exports every type and function the legacy callers
use (src/index.ts, src/tools/operator-doctor.ts, src/operator-doctor.test.ts,
src/server.test.ts resetOperatorDoctorCache,
src/tools/__tests__/operator-doctor.test.ts,
server/tools/dependency-arrival.test.ts) with the OLD call forms. It reads
process.env inside the call, not at module scope, which is what lets the
legacy tests keep flipping env per case and still see the change.

server/tools/operator-doctor.ts holds no config object, so it keeps importing
the src/ adapter rather than the server/ module; it repoints at L6 when the
adapter retires.

Behavior: unchanged except one latent gap the move closes. The old cache path
called buildOperatorDoctorStatus WITHOUT forwarding its options, so a caller's
qmdIndexPath or now() reached the builder only on the uncached direct call.
The options are now threaded through both.

Closure count: 39 -> 37 (the adapter is excluded by the M9 header regex).

Tests left in src/ (retire or move at L6): src/operator-doctor.test.ts,
src/server.test.ts, src/tools/__tests__/operator-doctor.test.ts -- untouched
per M1 EXCEPTION, passing through the adapter unchanged.
